+++ README.md
@@ -8,7 +8,7 @@ hond of kat te kiezen, passen wij deze concepten toe op het onderwerp Pokèmon!
 
 Mischien word je al nostalgisch als je het woord "pokemon" hoort en kan je nog steeds alle 152 originele pokemon uit je hoofd opnoemen in alfabetische volgorde, maar mischien ook niet. We zullen hier kort de belangrijkste aspecten van pokemon uitleggen die je nodig hebt voor de context van deze opdracht en de volgende.
 
-Pokemon is een japanse animatie serie. Op deze serie zijn inmiddels tal van computer en kaart spellen gebaseerd. In deze opdracht gaan we zo'n spel nabootsen. In de spellen is het zo dat jij, de speler, een pokemon trainer bent. Je doel is om pokemons te verzamelen en te trainen om uiteindelijk de beste pokemontrainer ter wereld te worden. Een pokemon is dus eigenlijk een soort dier, maar dan één die speciale krachten/aanvallen heeft. Er zijn een heleboel pokemon, hierboven zag je al dat er 152 originele pokemon zijn, maar inmiddels is daar al wel een vijfvoud van. Gelukkig zijn deze pokemon onder te verdelen in types, voor deze opdracht gebruiken we 4 verschillende types, maar er zijn er veel meer. Er zijn zelfs subtypes en nog veel meer variabelen, maar dat is voor deze opdrachten allemaal niet nodig.
+src.Pokemon is een japanse animatie serie. Op deze serie zijn inmiddels tal van computer en kaart spellen gebaseerd. In deze opdracht gaan we zo'n spel nabootsen. In de spellen is het zo dat jij, de speler, een pokemon trainer bent. Je doel is om pokemons te verzamelen en te trainen om uiteindelijk de beste pokemontrainer ter wereld te worden. Een pokemon is dus eigenlijk een soort dier, maar dan één die speciale krachten/aanvallen heeft. Er zijn een heleboel pokemon, hierboven zag je al dat er 152 originele pokemon zijn, maar inmiddels is daar al wel een vijfvoud van. Gelukkig zijn deze pokemon onder te verdelen in types, voor deze opdracht gebruiken we 4 verschillende types, maar er zijn er veel meer. Er zijn zelfs subtypes en nog veel meer variabelen, maar dat is voor deze opdrachten allemaal niet nodig.
 Een leuk feitje over pokemons is dat ze altijd één raar geluidje maken. Zo zegt een Pikachu "pika pika" en verder niks.
 
 ### Het spel element
@@ -30,7 +30,7 @@ Electric ⚡.
 Naast hun type, hebben Pokèmon natuurlijk bepaalde eigenschappen met elkaar gemeen. Daarom maken we gebruik van een
 Super klasse. 
 
-![Pokemon!](./Assets/pokemon.JPG)
+![src.Pokemon!](./Assets/pokemon.JPG)
 
 ## Randvoorwaarden
 
@@ -98,10 +98,10 @@ Het is uitdagender om jouw eigen stappenplan te maken, maar als je niet weet waa
 5. Maak de volgende subklassen aan en laat deze voldoen aan de gestelde randvoorwaarden (
    zie: [randvoorwaarden](#randvoorwaarden)):
 
-- `FirePokemon`
-- `WaterPokemon`
-- `GrassPokemon`
-- `ElectricPokemon`
+- `src.FirePokemon`
+- `src.WaterPokemon`
+- `src.GrassPokemon`
+- `src.ElectricPokemon`
 
 5. Instantieer in de `main` methode jouw 4 verschillende Pokèmon met de verschillende attributen.
